backend/requirements.txt
Django==5.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
faker==37.0.2
werkzeug==3.1.3
pytest==8.3.5
python-decouple==3.8
dj_database_url==2.3.0
dotenv==0.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework_simplejwt==5.5.0
faker-food==0.3.0
django_rest_swagger==2.2.0